Move stats into expr module
