+++ vortex-array/src/lib.rs
@@ -24,11 +24,11 @@ pub use context::*;
 pub use hash::*;
 pub use mask_future::*;
 pub use metadata::*;
+use session::rewrite::ArrayRewriteRuleRegistry;
+use vortex_session::registry::Registry;
 use vortex_session::Ref;
 use vortex_session::SessionExt;
-use vortex_session::registry::Registry;
 
-use crate::array::session::rewrite::ArrayRewriteRuleRegistry;
 use crate::array::transform::AnyArrayParent;
 use crate::array::transform::ArrayOptimizer;
 use crate::array::transform::ArrayParentReduceRule;
@@ -61,6 +61,7 @@ pub mod builders;
 mod canonical;
 pub mod compute;
 mod context;
+mod evaluate;
 pub mod execution;
 pub mod expr;
 mod hash;
@@ -73,6 +74,7 @@ pub mod patches;
 pub mod pipeline;
 pub mod search_sorted;
 pub mod serde;
+pub mod session;
 pub mod stats;
 pub mod stream;
 #[cfg(any(test, feature = "test-harness"))]
