+++ vortex-array/src/session/rewrite.rs
